+++ Bedtimestory/app/src/main/java/com/dragonlegend/kidstories/AddStoriesContentActivity.java
@@ -1,16 +1,29 @@
 package com.dragonlegend.kidstories;
 
+import android.app.Activity;
+import android.database.Cursor;
+import android.net.Uri;
+import android.provider.MediaStore;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.dragonlegend.kidstories.Utils.UploadImage;
+import com.pixplicity.easyprefs.library.Prefs;
+
 public class AddStoriesContentActivity extends AppCompatActivity implements View.OnClickListener {
     EditText mContentField;
     Button mSaveButton,mDiscardButton;
+    String name = "image";
+    private String title;
+    private String filePath;
+    private String content;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -28,7 +41,7 @@ public class AddStoriesContentActivity extends AppCompatActivity implements View
 //                finish();
 //            }
 //        });
-//        initViews();
+        initViews();
     }
 
     public void initViews(){
@@ -37,17 +50,41 @@ public class AddStoriesContentActivity extends AppCompatActivity implements View
         mSaveButton = findViewById(R.id.save_content);
         mDiscardButton.setOnClickListener(this);
         mSaveButton.setOnClickListener(this);
+
     }
+
+
 
     @Override
     public void onClick(View v) {
         switch (v.getId()){
             case R.id.save_content:
+                AddStory();
                 Toast.makeText(this,"Content Saved",Toast.LENGTH_SHORT).show();
                 break;
             default:
                 finish();
                 break;
         }
+    }
+
+    private void AddStory() {
+        content = mContentField.getText().toString().trim();
+        title = Prefs.getString("title", "");
+        String imageFileUri = Prefs.getString("filePathReal","");
+        Log.d("TAG", "AddStory: " + imageFileUri);
+        UploadImage.uploadFile(imageFileUri,name,title,content,this);
+    }
+
+    private String getImageAbsolutePath(Uri contentURI, Activity activity) {
+        Cursor cursor = activity.getContentResolver().query(contentURI, null, null, null, null);
+        if (cursor == null) {
+            return contentURI.getPath();
+        } else {
+            cursor.moveToFirst();
+            int index = cursor.getColumnIndex(MediaStore.Images.ImageColumns.DATA);
+            return cursor.getString(index);
+        }
+
     }
 }
