+++ Bedtimestory/app/src/main/java/com/dragonlegend/kidstories/Api/ApiInterface.java
@@ -7,12 +7,17 @@ import com.dragonlegend.kidstories.Api.Responses.StoryCategoryResponse;
 import com.dragonlegend.kidstories.Api.Responses.StoryResponse;
 import com.dragonlegend.kidstories.Model.User;
 
+import okhttp3.MultipartBody;
+import okhttp3.RequestBody;
+import okhttp3.ResponseBody;
 import retrofit2.Call;
 import retrofit2.http.Body;
 import retrofit2.http.Field;
 import retrofit2.http.FormUrlEncoded;
 import retrofit2.http.GET;
+import retrofit2.http.Multipart;
 import retrofit2.http.POST;
+import retrofit2.http.Part;
 import retrofit2.http.Path;
 
 public interface ApiInterface {
@@ -38,4 +43,13 @@ public interface ApiInterface {
 
     @GET("user/profile/{id}")
     Call<User> getProfile(@Path("id") String id);
+
+
+    @Multipart
+    @POST("story/create")
+    Call<ResponseBody> addStory(
+            @Part("title") RequestBody title,
+            @Part("story") RequestBody story,
+            @Part MultipartBody.Part image
+    );
 }
