--- Bedtimestory/app/src/main/java/com/example/bedtime/Api/Client.java
@@ -1,16 +1,0 @@
-package com.example.bedtime.Api;
-
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-public class Client {
-        public static final String BASE_URL = "https://dragon-legend-5.herokuapp.com/api/v1/";
-
-    public  static Retrofit getInstance(){
-        Retrofit retrofit = new Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build();
-        return retrofit;
-    }
-}
