--- Bedtimestory/app/src/main/java/com/example/bedtime/Api/Responses/BaseResponse.java
@@ -1,53 +1,0 @@
-package com.example.bedtime.Api.Responses;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class BaseResponse<T> {
-
-    @SerializedName("status")
-    @Expose
-    private Integer status;
-    @SerializedName("method")
-    @Expose
-    private String method;
-    @SerializedName("message")
-    @Expose
-    private String message;
-    @SerializedName("data")
-    @Expose
-    private T data;
-
-    public Integer getStatus() {
-        return status;
-    }
-
-    public void setStatus(Integer status) {
-        this.status = status;
-    }
-
-    public String getMethod() {
-        return method;
-    }
-
-    public void setMethod(String method) {
-        this.method = method;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public T getData() {
-        return data;
-    }
-
-    public void setData(T data) {
-        this.data = data;
-    }
-
-}
