--- Bedtimestory/app/src/main/java/com/example/bedtime/Api/Responses/CategoryAllResponse.java
@@ -1,54 +1,0 @@
-package com.example.bedtime.Api.Responses;
-
-import com.example.bedtime.Model.Category;
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class CategoryAllResponse {
-    @SerializedName("status")
-    @Expose
-    private Integer status;
-    @SerializedName("method")
-    @Expose
-    private String method;
-    @SerializedName("message")
-    @Expose
-    private String message;
-    @SerializedName("data")
-    @Expose
-    private List<Category> data = null;
-
-    public Integer getStatus() {
-        return status;
-    }
-
-    public void setStatus(Integer status) {
-        this.status = status;
-    }
-
-    public String getMethod() {
-        return method;
-    }
-
-    public void setMethod(String method) {
-        this.method = method;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public List<Category> getData() {
-        return data;
-    }
-
-    public void setData(List<Category> data) {
-        this.data = data;
-    }
-}
