--- Bedtimestory/app/src/main/java/com/example/bedtime/Api/Responses/LoginResponse.java
@@ -1,36 +1,0 @@
-package com.example.bedtime.Api.Responses;
-
-import com.example.bedtime.Model.User;
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class LoginResponse {
-    @SerializedName("status")
-    @Expose
-    private Integer status;
-    @SerializedName("method")
-    @Expose
-    private String method;
-    @SerializedName("message")
-    @Expose
-    private String message;
-    @SerializedName("data")
-    @Expose
-    private User user;
-
-    public Integer getStatus() {
-        return status;
-    }
-
-    public String getMethod() {
-        return method;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public User getUser() {
-        return user;
-    }
-}
