--- Bedtimestory/app/src/main/java/com/example/bedtime/Api/Responses/StoryReactionResponse.java
@@ -1,154 +1,0 @@
-package com.example.bedtime.Api.Responses;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class StoryReactionResponse {
-    @SerializedName("cat_id")
-    @Expose
-    private List<String> catId = null;
-    @SerializedName("_id")
-    @Expose
-    private String id;
-    @SerializedName("release_date")
-    @Expose
-    private String releaseDate;
-    @SerializedName("comments")
-    @Expose
-    private List<Object> comments = null;
-    @SerializedName("title")
-    @Expose
-    private String title;
-    @SerializedName("story")
-    @Expose
-    private String story;
-    @SerializedName("designation")
-    @Expose
-    private String designation;
-    @SerializedName("author")
-    @Expose
-    private String author;
-    @SerializedName("likes")
-    @Expose
-    private Integer likes;
-    @SerializedName("dislikes")
-    @Expose
-    private Integer dislikes;
-    @SerializedName("imageId")
-    @Expose
-    private String imageId;
-    @SerializedName("image")
-    @Expose
-    private String image;
-    @SerializedName("__v")
-    @Expose
-    private Integer v;
-
-    public List<String> getCatId() {
-        return catId;
-    }
-
-    public void setCatId(List<String> catId) {
-        this.catId = catId;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getReleaseDate() {
-        return releaseDate;
-    }
-
-    public void setReleaseDate(String releaseDate) {
-        this.releaseDate = releaseDate;
-    }
-
-    public List<Object> getComments() {
-        return comments;
-    }
-
-    public void setComments(List<Object> comments) {
-        this.comments = comments;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getStory() {
-        return story;
-    }
-
-    public void setStory(String story) {
-        this.story = story;
-    }
-
-    public String getDesignation() {
-        return designation;
-    }
-
-    public void setDesignation(String designation) {
-        this.designation = designation;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public Integer getLikes() {
-        return likes;
-    }
-
-    public void setLikes(Integer likes) {
-        this.likes = likes;
-    }
-
-    public Integer getDislikes() {
-        return dislikes;
-    }
-
-    public void setDislikes(Integer dislikes) {
-        this.dislikes = dislikes;
-    }
-
-    public String getImageId() {
-        return imageId;
-    }
-
-    public void setImageId(String imageId) {
-        this.imageId = imageId;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    public Integer getV() {
-        return v;
-    }
-
-    public void setV(Integer v) {
-        this.v = v;
-    }
-
-}
-
