--- Bedtimestory/app/src/main/java/com/example/bedtime/Config.java
@@ -1,5 +1,0 @@
-package com.example.bedtime;
-
-public class Config {
-    public static final String USER_ID = "user_id";
-}
