--- Bedtimestory/app/src/main/java/com/example/bedtime/Model/Category.java
@@ -1,45 +1,0 @@
-package com.example.bedtime.Model;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Category {
-
-    @SerializedName("_id")
-    @Expose
-    protected String id;
-    @SerializedName("name")
-    @Expose
-    private String name;
-    @SerializedName("image")
-    @Expose
-    private String image;
-
-    public Category(String name) {
-        this.name = name;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-}
