--- Bedtimestory/app/src/main/java/com/example/bedtime/Model/Story.java
@@ -1,143 +1,0 @@
-
-package com.example.bedtime.Model;
-
-import java.util.List;
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Story {
-
-    @SerializedName("cat_id")
-    @Expose
-    private List<Category> mCategories = null;
-    @SerializedName("_id")
-    @Expose
-    private String id;
-    @SerializedName("release_date")
-    @Expose
-    private String releaseDate;
-    @SerializedName("likes")
-    @Expose
-    private int likes ;
-    @SerializedName("comments")
-    @Expose
-    private List<Object> comments = null;
-    @SerializedName("title")
-    @Expose
-    private String title;
-    @SerializedName("story")
-    @Expose
-    private String story;
-    @SerializedName("designation")
-    @Expose
-    private String designation;
-    @SerializedName("author")
-    @Expose
-    private String author;
-    @SerializedName("imageId")
-    @Expose
-    private String imageId;
-    @SerializedName("image")
-    @Expose
-    private String image;
-    @SerializedName("__v")
-    @Expose
-    private Integer v;
-
-    public List<Category> getCategories() {
-        return mCategories;
-    }
-
-    public void setCategories(List<Category> categories) {
-        this.mCategories = categories;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getReleaseDate() {
-        return releaseDate;
-    }
-
-    public void setReleaseDate(String releaseDate) {
-        this.releaseDate = releaseDate;
-    }
-
-    public int getLikes() {
-        return likes;
-    }
-
-    public void setLikes(int likes) {
-        this.likes = likes;
-    }
-
-    public List<Object> getComments() {
-        return comments;
-    }
-
-    public void setComments(List<Object> comments) {
-        this.comments = comments;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getStory() {
-        return story;
-    }
-
-    public void setStory(String story) {
-        this.story = story;
-    }
-
-    public String getDesignation() {
-        return designation;
-    }
-
-    public void setDesignation(String designation) {
-        this.designation = designation;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public String getImageId() {
-        return imageId;
-    }
-
-    public void setImageId(String imageId) {
-        this.imageId = imageId;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    public Integer getV() {
-        return v;
-    }
-
-    public void setV(Integer v) {
-        this.v = v;
-    }
-
-}
