--- Bedtimestory/app/src/main/java/com/example/bedtime/Model/User.java
@@ -1,137 +1,0 @@
-package com.example.bedtime.Model;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class User {
-    @SerializedName("token")
-    @Expose
-    private String token;
-    @SerializedName("id")
-    @Expose
-    private String id;
-    @SerializedName("name")
-    @Expose
-    private String name;
-    @SerializedName("email")
-    @Expose
-    private String email;
-    @SerializedName("admin")
-    @Expose
-    private Boolean admin;
-    @SerializedName("premium")
-    @Expose
-    private Boolean premium;
-    @SerializedName("bookmark")
-    @Expose
-    private List<Object> bookmark = null;
-    @SerializedName("bookmark_count")
-    @Expose
-    private Integer bookmarkCount;
-    @SerializedName("liked")
-    @Expose
-    private Integer liked;
-    @SerializedName("image")
-    @Expose
-    private String image;
-    private String designation;
-    private String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public Boolean getAdmin() {
-        return admin;
-    }
-
-    public void setAdmin(Boolean admin) {
-        this.admin = admin;
-    }
-
-    public Boolean getPremium() {
-        return premium;
-    }
-
-    public void setPremium(Boolean premium) {
-        this.premium = premium;
-    }
-
-    public List<Object> getBookmark() {
-        return bookmark;
-    }
-
-    public void setBookmark(List<Object> bookmark) {
-        this.bookmark = bookmark;
-    }
-
-    public Integer getBookmarkCount() {
-        return bookmarkCount;
-    }
-
-    public void setBookmarkCount(Integer bookmarkCount) {
-        this.bookmarkCount = bookmarkCount;
-    }
-
-    public Integer getLiked() {
-        return liked;
-    }
-
-    public void setLiked(Integer liked) {
-        this.liked = liked;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDesignation() {
-        return designation;
-    }
-
-    public void setDesignation(String designation) {
-        this.designation = designation;
-    }
-}
