+++ Bedtimestory/app/src/main/res/values/strings.xml
@@ -44,6 +44,7 @@
     <string name="discard">Discard</string>
     <string name="content">Content</string>
     <string name="phone_number">Phone Number</string>
+    <string name="internal_storage">internal storage</string>
     <string-array name="designation_array">
         <item>Kid</item>
         <item>Adult</item>
